Fixes #9582: display content tab after content view is loaded BZ1196720.
We were displaying the yum content tab before we knew whether ot not
the content view was a composite.  In extreme cases the tab was around
long enough for the user to click on it.  This commit does not show the
tab until the content view has loaded and we can show the correct menu.

http://projects.theforeman.org/issues/9582
https://bugzilla.redhat.com/show_bug.cgi?id=1196720
(cherry picked from commit 0e91e0d096725a0d7443d6a62b15110c15541adb)