Fixes #9583: improve repository deletion responses BZ 1166365.
This commit fixes a few issues around repository removal.

When deleting a single repository we were spawning an async task
but not showing the task.  With this commit removing a repository
was changed to be a synchronous task.

When deleting multiple repositories we were creating several async
tasks but not returning any task IDs.  This commit instead creates
a parent task and returns that task to the user.

Also fixed is a ISE caused by referencing [] on nil pulp repo facts.

http://projects.theforeman.org/issues/9583
https://bugzilla.redhat.com/show_bug.cgi?id=1166365