Fixes #9934: do not refresh errata on initial load BZ 1206704.
We were loading the errata nutupane twice which was causing an
issue where infinite scroll was not fetching enough results to
scroll on certain resolutions.  This commit only reloads the
nutupane if it's not the initial load.

http://projects.theforeman.org/issues/9934
https://bugzilla.redhat.com/show_bug.cgi?id=1206704
(cherry picked from commit 00453e99ada32bf73ee6bfd25bc9f46f2733b684)