Fixes #9915: Backend cleanup script handles uebercert and hypervisors.
When an uebercert is generated Candlepin generates a consumer. If
a user runs the cleanup script we don't want to attempt to delete
as it leads to errors and is not a consumer we manage directly.
Also, we previously were preventing deletion of Hypervisors due
in part to the errors that the script would generate since Hypervisors
do not create corresponding Pulp consumers.

(cherry picked from commit d2cc576e48039d36723ed55d6e4bfcc1838a252e)