Fixes #10620 - Repos no longer protected by default
Repositories created via UI are not protected by default, however this
is not the case for CLI. Repos created by the cli tend to have the
protection on by default. This does not work well for docker, so trying
to be consistent here.
Also updated code to always set protected to false for docker type. This
is because docker type does not have any protection at this time.

more

(cherry picked from commit 70dda2a57fcf4f285b6952d073f93842331622ed)