Fixes #10681:  show correct installable count BZ 1226997.
The installable count was incorrect when a repository was
provided becuase the query was incorrect.  This commit
fixes the query for systems when a repository is provided.
In addition, this commit prevents only installable from
being selected (as installable is a subset of applicable)
and shows a tooltip in that case.

http://projects.theforeman.org/issues/10681
https://bugzilla.redhat.com/show_bug.cgi?id=1226997