Fixes #10697 - Improved CP SetContent add/delete logic
When setting/deleting the content id for an environment in candlepin a race
condition can happen. This because the SetContent action says:

1) Fetch me the list of existing content ids for this environment from
candlepin.
2) Take the existing content ids list in step 1 and subtract that from the
list of content ids provided in the input. In other words give me the list
of content ids to add.
3) Add the new content ids to environment in candlepin.

Problem is between step 1 and step 3 there can be a race condition
if 2 repos belonging to the same environment are clicked quickly, because
they run in parallel. It is possible that the content id has already
been added by one of the repo tasks causing a dupe error when the 2 nd
repo action tries to run.

When this problem occured candlepin used to throw 500. That part of the
code has now been fixed and instead a 409 is raised. This commit handles
the 409 instead of 500.

Also this commit handles the delete ids case. Again the problem here is
1) Fetch the list of existing content ids for this environment from
candlepin.
2) Take the list of content ids provided by the input, subtract that
from the existing content ids (obtained in step 1.) In other words get
the list of content ids to delete.
3) Remove  the new content ids from the environment in candlepin.

Problem is between step 1 and step 3 there can be a race condition
if 2 repos belonging to the same environment are clicked quickly,
because they run in parallel. It is possible that the content id has
already been deleted by one of the repo actions there by causing a 404
when the second repo action runs.

Both these conditions should now be fixed with this commit.

Still have the retry logic because we do not want to go to an infinite
loop if Candlepin returns a 404 for not loading properly.

(cherry picked from commit c2aa177c3ae76a02bfaa1071a65628e32e332626)