Fixes BZ{1198349,1222365,1222673,1228278,1232510}: fix errata count.
There were several places in the UI where the errata count was
off due to downstream using an old version of activerecord with
a bug in it.  Upstream we are using 3.2.21 but downstream we are
stuck with 3.2.8.  The activerecord bug was fixed in 3.2.13.  This
commit monkey patches 3.2.8 with the fix.

https://bugzilla.redhat.com/show_bug.cgi?id=1198349
https://bugzilla.redhat.com/show_bug.cgi?id=1222365
https://bugzilla.redhat.com/show_bug.cgi?id=1222673
https://bugzilla.redhat.com/show_bug.cgi?id=1228278
https://bugzilla.redhat.com/show_bug.cgi?id=1232510

See also https://github.com/rails/rails/issues/6865