BZ #1507498 - Speed up clean backend objects
The clean backend objects rake task would loop through every host and
figure out if the candlepin and pulp ids were valid. This meant for
every host there would be 2 rest calls for this check. On machines where
the number of hosts were in the 10 s of thousands this would mean 20K
rest calls.
On top of that the final clean "orphaned" systems, i.e.'find all systems
in candlepin that are not in katello' would potentially load up or
serialize all candlepin consumers in 1 very big json (very expensive
task) to determine if there are orphaned consumers.

This commit tries to address this load by
1) Reducing the json foot print in candlepin (just get the UUIDs and
nothing else)
2) Page the calls. (Fetch 5000 uuids at a time, dont get everything)
3) Common sense arithmetic to calculate/fetch hosts that are in Katello
but not in CP/Pulp and vice versa
4) Hopefully more modular code for readability.