BZ #1522912 - use archive repo for CVV env indexing
this commit changes the way to we index repositories
that are published in a content view within the lifecycle
environment.  Previously we went to pulp to fetch the uuids.
Now we base it off the "archive" copy of that repository.

This may be slightly faster (or much faster if pulp/mongo
is under heavy load), but will facilitate even faster publishes
in the future by not copying any actual units to the repo.