BZ #1513769 - CV Docker Filters selects correctly
Consider Content View with a Docker repo and Docker Tag filter that
says filter by Tag "goo". And assume that the same Manifest exists in
two separate repositories Repo1 and Repo2 where
Repo1 =>Manifest1=>Tags [latest, goo] and
Repo2 =>Manifest1=> Tags[latest]
Now suppose the ContentView had Repo2 along with the Tag Filter to
include manifests with the name 'goo' and wanted to publish it. Our
filter should copy no Tags because there is nothing called 'goo' in
repo2.
Before this commit however this code would include goo in its clause
because of a faulty query.
This commit fixes that by improving the query to include tags in the
repository belonging to the manifest first.

This commit also fixes a "docker tag counts" issue in the content view
version so that tags with the same name for schema1 and 2 are
not double counted.

Finally it removes an unused method in the content view version also
related to docker_tags

(cherry picked from commit 9b45392b655ff0c6ab81b800b7e6007d3faafa03)